+++ server/src/main/java/com/inventage/portal/gateway/proxy/middleware/responseSessionCookie/ResponseSessionCookieRemovalMiddleware.java
@@ -27,7 +27,7 @@ public class ResponseSessionCookieRemovalMiddleware implements Middleware {
 
     public ResponseSessionCookieRemovalMiddleware(String name, String sessionCookieName) {
         this.name = name;
-        this.sessionCookieName = (sessionCookieName == null) ? SessionMiddleware.COOKIE_NAME_DEFAULT
+        this.sessionCookieName = (sessionCookieName == null) ? SessionMiddleware.SESSION_COOKIE_NAME_DEFAULT
                 : sessionCookieName;
     }