+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/sessionBag/SessionBagMiddlewareTest.java
@@ -261,7 +261,7 @@ public class SessionBagMiddlewareTest {
             // setup proxy
             SessionHandler sessionHandler = SessionHandler.create(sessionStore).setSessionCookieName(sessionCookieName);
             SessionBagMiddleware sessionBag = new SessionBagMiddleware("sessionBag", whitelistedCookies,
-                    "inventage-portal-gateway.session");
+                    "uniport.session");
             ProxyMiddleware proxy = new ProxyMiddleware(vertx, "proxy", host, servicePort);
             Router proxyRouter = Router.router(vertx);
             proxyRouter.route().handler(sessionHandler).handler(sessionBag).handler(proxy);